fixup test

+++ test/conftest.py
@@ -18,7 +18,9 @@ def dummy_data() -> miniDAS:
     start_time = int(datetime.utcnow().timestamp() * 1e9)
 
     rstate = np.random.RandomState(123123)
-    traces = rstate.randint(low=-10000, high=10000, size=(nchannels, nsamples))
+    traces = rstate.randint(
+        low=-10000, high=10000, size=(nchannels, nsamples), dtype=np.int32
+    )
 
     lat0 = 48.858  # Eiffel Tower
     lon0 = 2.2945
